requirements.txt
# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.2

# OpenAI
openai>=1.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.3.1
httpx>=0.24.0
pytest-cov>=4.1.0 